% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.scatter.R
\name{gam.scatter}
\alias{gam.scatter}
\title{GAM version of smooth.scatter()}
\usage{
gam.scatter(x, y, scatter = FALSE, quant.lines = NULL)
}
\arguments{
\item{x, y}{numeric vectors to be included in the scatterplot}

\item{scatter, }{logical, if TRUE plots the scatterplot, if FALSE only the GAM line}
}
\value{
invisible output of GAM model and plotted lines
}
\description{
Scatter plot between x and y, with best fitting GAM curve.
}
\examples{
\dontrun{
x = rnorm(500)
y = x^2+rnorm(500)
gam.scatter(x,y) 
}
}
