% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.plot.R
\name{prepare.plot}
\alias{prepare.plot}
\title{Start a pre-formatted Base-R plot}
\usage{
prepare.plot(
  vars = list(),
  ylim.padding = c(0, 0),
  longest.label = 2,
  x1.lab = "x1.lab",
  x2.lab = "(x2.lab)",
  y1.lab = "y1.lab",
  y2.lab = "(y2.lab)",
  main1 = "main1",
  main2 = "(main2)"
)
}
\arguments{
\item{vars}{list of numeric vectors with all variables that will be plotted.
The length of the first variable determines the x-axis size. e.g., if it has 20
numbers, the x coordinates will be 1-20}

\item{ylim.padding}{numeric vector of length two, providing the percentage
padding to be added to the bottom and top of the automatically set ylim.}

\item{longest.label}{either numeric scalar, or character vector If numeric, it
is the length of the longest label to use for a x-axis tickmark. If character
vector it is the set of labels to use, and the longest length is computed based
on it.}

\item{x1.lab}{character string, 1st line label for x-axis, shown in bold.}

\item{x2.lab}{character string, 2nd line label for x-axis, shown in italics.}

\item{y1.lab}{character string, 1st line label for y-axis, shown in bold.}

\item{y2.lab}{character string, 2nd line label for y-axis, shown in italics.}

\item{main1.lab}{character string, 1st line label for figure header, shown in bold.}

\item{main2.lab}{character string, 2nd line label for figure header, shown in italics.}
}
\value{
NULL, for it produces a figure, and does not conduct calculations.
}
\description{
The x and y axes, and the figure main headers, are  described with 1 line in
bold, and 1 in italics below it (optional). If labels for the x-axis tick marks
are provided, the x-axis automatically gets padding to fit the text.
The ylim argument is automatically set to fit all variables and can be expanded
with the \code{ylim.padding} argument.
}
\examples{
\dontrun{
x1=rnorm(25)
x2=x1+1
x3=x1-1
prepare.plot(list(x1,x2,x3))
points(x1)
points(x2)

prepare.plot(list(x1,x2,x3),x1.lab='These are the 25 objects')
points(x3,type='l')

}

}
